requirements.txt
Flask==1.1.2
pymongo==3.11.2
requests==2.25.1
python-dateutil==2.8.1
cbpro==1.1.4